+++ glean/src/core/utils.ts
@@ -75,7 +75,7 @@ export function isUndefined(v: unknown): v is undefined {
  *          stating whether `v` is a string.
  */
 export function isString(v: unknown): v is string {
-  return (typeof v === "string" || (typeof v === "object" && v !== null && v.constructor === String));
+  return typeof v === "string";
 }
 
 /**
@@ -87,7 +87,7 @@ export function isString(v: unknown): v is string {
  *          stating whether `v` is a boolean.
  */
 export function isBoolean(v: unknown): v is boolean {
-  return (typeof v === "boolean" || (typeof v === "object" && v !== null && v.constructor === Boolean));
+  return typeof v === "boolean";
 }
 
 /**
@@ -99,7 +99,19 @@ export function isBoolean(v: unknown): v is boolean {
  *          stating whether `v` is a number.
  */
 export function isNumber(v: unknown): v is number {
-  return ((typeof v === "number" || (typeof v === "object" && v !== null && v.constructor === Number)) && !isNaN(v));
+  return typeof v === "number" && !isNaN(v);
+}
+
+/**
+ * Checks whether or not `v` is an integer.
+ *
+ * @param v The value to verify.
+ *
+ * @returns A special Typescript value (which compiles down to a boolean)
+ *          stating whether `v` is a number.
+ */
+export function isInteger(v: unknown): v is number {
+  return isNumber(v) && Number.isInteger(v);
 }
 
 /**